Fix: Remove ACL and enforce bucket ownership for objects (#18)
Access Control Lists (ACLs) are not best-practice for S3 and should be
removed. We use a bucket policy instead.
Also, the recommended and best-practice setting for a S3 bucket is to
enforce bucket ownership. This means all objects uploaded to a bucket
(from whatever account) are owned by the account owning the bucket. This
configuration is explicitly set now. With this configuration enabled,
ACLs are disabled automatically.