Add Terraform modules for API Gateway endpoints (#64)
- Add Terraform modules for API Gateway endpoints to reduce code
redundancy
- Specify distinct version for Terraform linter (tflint): avoid pipeline
errors by automatic version upgrades
- Simplify update skill function for DynamoDB: update complete skill
attribute instead of specific values

NOTICE:
tflint does not work for the TF files in the module folders.
This is because it emits warnings for missing providers and versions
although these are defined in the versions.tf in the terraform folder.
However, tflint seems to ignore this file when checking the module
folders. No option worked to check the subfolders with --recursive and
at the same time avoid the warnings. Even ignoring the warnings in the
tflint config file didn't work.