fix: Make patch combat values endpoint idempotent again (#102)
- fix: patch combat values endpoint does not update the combat values
anymore for multiple calls with the same payload (idempotent behavior)
- fix: Split tests for update combat values into idempotency and update
test cases. The error of non idempotent behavior for the patch combat
values endpoint was not detected because a) the test data values were
hardcoded and b) the test assertions were not correct for the
idempotency tests
- Update tests for idempotency: Reference values from test data and
avoid hardcoded values. This should avoid undetected errors as for the
'update combat values' case.