[BREAKING CHANGES] Include combat stats in patch-base-value endpoint (#104)
**BREAKING CHANGES:**
- Add effect on combat stats to patch-base-value endpoint -> changed API
schema
- Add logic to include combat stats in changed base value when adding
and reverting history records

Misc:
- fix: Use schema of level and special abilities change for:
  - the API responses of the corresponding endpoints
- schema for the corresponding history record entries (adding and
reverting)
- fix: Update response of patch-attribute endpoint to include only
changed combat stats for the 'old' data object. The 'new' data object
already included only changed combat stats. Now 'old' and 'new' are
consistent in terms of the returned data structure. The API schema of
the response is NOT affected by this (non-breaking fix).
- test: Update test record for ATTRIBUTE_CHANGED (with base value change
but without combat stats change) -> make it more consistent with the
actual ruleset to avoid confusion when reading the tests: Use attribute
in test data that does not include changed combat stats